Page level scrubbable numbers
Thanks to @david-trumid for the start on this. Continued to convert
Numbers component into a full class at NumberScrubber.jsx.
